Organize files related to TextSpanNotifier in a folder
