+++ lib/src/text_span/gestures.dart
