+++ lib/src/text_span/transient_elements_builder.dart
